ColorMystery/requirements.txt
numpy
opencv-python-headless
Pillow
scikit-learn
cairosvg
streamlit
pytest
flake8
black
